+++ app/views/events/show.html.erb
@@ -45,7 +45,7 @@
     </div>
     <!-- Action Buttons -->
     <div class="event-actions text-center mt-3">
-      <% if (@event.early_access_for_members? && current_user&.member?) || @event.general_registration_open? %>
+      <% if (@event.early_access_for_members? && current_user&.membership_active?) || @event.general_registration_open? %>
         <%= button_to 'Register Now', event_payments_path(@event, paymentable_type: 'event', paymentable_id: @event.id), method: :post, class: 'event-payments btn btn-primary mb-3', data: { turbo: false } %>
       <% end %>
       <% if current_user&.role_id == 1 %>
