+++ app/views/pages/about.html.erb
@@ -1,5 +1,5 @@
 <% content_for :title, "About Us" %>
-<h1 >About Us</h1>
+<h1 class="text-center mt-3">About Us</h1>
 <div class="container my-5 about">
   <div class="row">
     <div class="col-lg-8 offset-lg-2">
